Move OpenStack stuff to its own module and avoid importing it except on Linux.
