+++ flocker/common/__init__.py
@@ -9,6 +9,14 @@ __all__ = [
     'auto_threaded', 'auto_openstack_logging',
 ]
 
+import platform
+
 from ._ipc import INode, FakeNode, ProcessNode
 from ._defer import gather_deferreds
-from ._thread import auto_threaded, auto_openstack_logging
+from ._thread import auto_threaded
+
+
+if platform().system == 'Linux':
+    # For some reason I don't understand,  keystoneclient has problems on OS X.
+    # Fortunately, we don't need keystoneclient on OS X.
+    from ._openstack import auto_openstack_logging
